[feat] #39 키보드 내리기 기능 extension 분리 및 추가 적용
