+++ SniffMeet/SniffMeet/Source/Common/Extension + Keyboard.swift
@@ -1,0 +1,21 @@
+//
+//  Extension + Keyboard.swift
+//  SniffMeet
+//
+//  Created by 배현진 on 11/12/24.
+//
+
+import UIKit
+
+extension UIViewController {
+    func hideKeyboardWhenTappedAround() {
+        let tap = UITapGestureRecognizer(target: self,
+                                         action: #selector(ProfileInputView.dismissKeyboard))
+        tap.cancelsTouchesInView = false
+        view.addGestureRecognizer(tap)
+    }
+
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+}
